+++ CHANGES.rst
@@ -5,7 +5,7 @@ Change log for sw.allotmentclub
 5.7.3 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Bugfix: Calculation of energy value with advanced pay.
 
 
 5.7.2 (2022-08-31)